change the datatype

+++ file.py
@@ -1,8 +1,8 @@
 
-num1 = float(input("Enter the your first number: "))
-num2 = float(input("Enter the  your second number: "))
-num3 = float(input("Enter the your  third number: "))
-num4 = float(input("Enter the  your fourth number: "))
+num1 = int(input("Enter the your first number: "))
+num2 = int(input("Enter the  your second number: "))
+num3 = int(input("Enter the your  third number: "))
+num4 = int(input("Enter the  your fourth number: "))
 
 mul_result = num1 * num2 * num3 * num4
 print(f"The sum of the four numbers is: {mul_result}")
